requirements.txt
backgroundremover
certifi
charset-normalizer
commandlines
decorator
ffmpeg-python
filelock
filetype
fsspec
future
hsh
idna
ImageHash
imageio
imageio-ffmpeg
Jinja2
lazy_loader
llvmlite
MarkupSafe
more-itertools
moviepy
mpmath
networkx
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python
packaging
Pillow
proglog
PyMatting
PySocks
PyWavelets
requests
scikit-image
scipy
setuptools
six
sympy
tifffile
torch
torchvision
tqdm
triton
typing_extensions
urllib3
waitress
